requirements.txt
# activate
# annotated-types
# anyio
# asgiref
# certifi
# charset-normalizer
# colorama
# contourpy
# crispy-bootstrap5
# cycler
# diskcache
# distlib
# distro
# dj-database-url
# Django
# django-cors-headers
# django-crispy-forms
# django-environ
# django-filter
# django-pwdprotect
# django-tokenauth
# djangorestframework
# docker
# filelock
# FLAML
# fonttools
# h11
# httpcore
# httpx
# idna
# kiwisolver
# Markdown
# matplotlib
# migrations
# MouseInfo
# mpmath
# numpy
# openai
# packaging
# Pillow
# platformdirs
# psycopg2-binary
# pyautogen
# PyAutoGUI
# pydantic
# pydantic_core
# PyGetWindow
# Pygments
# PyMsgBox
# pyparsing
# pyperclip
# PyRect
# PyScreeze
# python-dateutil
# python-dotenv
# pytweening
# pytz
# regex
# requests
# scipy
# six
# sniffio
# sqlparse
# sympy
# termcolor
# tiktoken
# tokenauth
# tqdm
# typing_extensions
# tzdata
# urllib3
# virtualenv
# virtualenvwrapper-win







#####




aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
infobip-api-python-client==3.0.3
infobip-api-python-sdk==5.0.1
Markdown==3.6
multidict==6.0.5
packaging==24.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
pydantic==1.10.15
PyJWT==2.8.0
python-dateutil==2.9.0.post0
regex==2022.3.15
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==9.0.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
yarl==1.9.4



